# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_handler.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  secret-service_internal_domain.ServiceAccount:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      status:
        $ref: '#/definitions/secret-service_internal_domain.ServiceAccountStatus'
      updated_at:
        type: string
    type: object
  secret-service_internal_domain.ServiceAccountStatus:
    enum:
    - active
    - revoked
    type: string
    x-enum-varnames:
    - ServiceAccountStatusActive
    - ServiceAccountStatusRevoked
  secret-service_internal_domain.Team:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  secret-service_internal_domain.TeamMember:
    properties:
      created_at:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/secret-service_internal_domain.TeamRole'
      team_id:
        type: string
      user_id:
        type: string
    type: object
  secret-service_internal_domain.TeamRole:
    enum:
    - owner
    - member
    type: string
    x-enum-varnames:
    - TeamRoleOwner
    - TeamRoleMember
...
  secret-service_internal_dto.AddTeamMemberRequest:
    properties:
      role:
        type: string
      user_id:
        type: string
    type: object
...
  secret-service_internal_dto.CreateServiceAccountRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  secret-service_internal_dto.CreateServiceAccountResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      token:
        description: shown only on creation
        type: string
      warning:
        type: string
    type: object
  secret-service_internal_dto.CreateTeamRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  secret-service_internal_dto.ServiceAccountLoginRequest:
    properties:
      service_account_id:
        type: string
      token:
        type: string
    type: object
...
  secret-service_internal_dto.TokenResponse:
    properties:
      access_token:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Centralized API key storage and management service for development
    teams.
  title: Secret Service API
  version: "1.0"
paths:
...
  /auth/service-login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Service account ID and token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.ServiceAccountLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/secret-service_internal_dto.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      summary: Authenticate service account
      tags:
      - service-accounts
...
  /projects/{projectID}/service-accounts:
    get:
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret-service_internal_domain.ServiceAccount'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List project service accounts
      tags:
      - service-accounts
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: projectID
        required: true
        type: string
      - description: Service account name and description
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.CreateServiceAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/secret-service_internal_dto.CreateServiceAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create service account
      tags:
      - service-accounts
...
  /service-accounts/{saID}/revoke:
    post:
      parameters:
      - description: Service account ID
        in: path
        name: saID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Revoke service account
      tags:
      - service-accounts
  /teams:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret-service_internal_domain.Team'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List user teams
      tags:
      - teams
    post:
      consumes:
      - application/json
      parameters:
      - description: Team name and description
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/secret-service_internal_domain.Team'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create team
      tags:
      - teams
  /teams/{teamID}/members:
    get:
      parameters:
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/secret-service_internal_domain.TeamMember'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: List team members
      tags:
      - teams
    post:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      - description: User ID and role (owner/member)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/secret-service_internal_dto.AddTeamMemberRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Add member to team
      tags:
      - teams
  /teams/{teamID}/members/{userID}:
    delete:
      parameters:
      - description: Team ID
        in: path
        name: teamID
        required: true
        type: string
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Remove member from team
      tags:
      - teams
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"